move functions into new file
